gitlab_remove_doubleton_members/requirements.txt
python-gitlab==4.11.1
urllib3==2.2.3
